requirements.txt
Django==1.11.4
pytz==2017.2
djangorestframework==3.6.3
django-bulk-update==2.1.0
psycopg2